Fix bug with decimals and multiple operators
If a user pressed an operation button multiple times in a row, the
calculator would break. Now, we check to see if the last input was an
operator and, if so, ignore the button press. In the future, maybe we
will replace the current  operation with the most recently pressed
operator.

Previously, a user could enter as many decimals as they wanted, which
obviously breaks the calculator. Now, we check if there is a decimal
already and, if so, ignore any further decimals.